Creating a CitiesContext
