+++ src/App.jsx
@@ -1,4 +1,3 @@
-import { useState, useEffect } from "react";
 import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
 
 import Product from "./pages/Product";
@@ -11,59 +10,31 @@ import CityList from "./components/CityList";
 import CountryList from "./components/CountryList";
 import City from "./components/City";
 import Form from "./components/Form";
-
-const BASE_URL = "http://localhost:8000";
+import { CitiesProvider } from "./contexts/CitiesContext";
 
 function App() {
-	const [cities, setCities] = useState([]);
-	const [isLoading, setIsLoading] = useState(false);
-
-	useEffect(function () {
-		async function fetchCities() {
-			try {
-				setIsLoading(true);
-				const res = await fetch(`${BASE_URL}/cities`);
-				const data = await res.json();
-				setCities(data);
-			} catch {
-				alert("There was an error loading data...");
-			} finally {
-				setIsLoading(false);
-			}
-		}
-		fetchCities();
-	}, []);
-
 	return (
-		<BrowserRouter>
-			<Routes>
-				<Route index element={<Homepage />} />
-				<Route path="product" element={<Product />} />
-				<Route path="pricing" element={<Pricing />} />
-				<Route path="login" element={<Login />} />
-				<Route path="app" element={<AppLayout />}>
-					<Route index element={<Navigate replace to="cities" />} />
-					<Route
-						path="cities"
-						element={
-							<CityList cities={cities} isLoading={isLoading} />
-						}
-					/>
-					<Route path="cities/:id" element={<City />} />
-					<Route
-						path="countries"
-						element={
-							<CountryList
-								cities={cities}
-								isLoading={isLoading}
-							/>
-						}
-					/>
-					<Route path="form" element={<Form />} />
-				</Route>
-				<Route path="*" element={<PageNotFound />} />
-			</Routes>
-		</BrowserRouter>
+		<CitiesProvider>
+			<BrowserRouter>
+				<Routes>
+					<Route index element={<Homepage />} />
+					<Route path="product" element={<Product />} />
+					<Route path="pricing" element={<Pricing />} />
+					<Route path="login" element={<Login />} />
+					<Route path="app" element={<AppLayout />}>
+						<Route
+							index
+							element={<Navigate replace to="cities" />}
+						/>
+						<Route path="cities" element={<CityList />} />
+						<Route path="cities/:id" element={<City />} />
+						<Route path="countries" element={<CountryList />} />
+						<Route path="form" element={<Form />} />
+					</Route>
+					<Route path="*" element={<PageNotFound />} />
+				</Routes>
+			</BrowserRouter>
+		</CitiesProvider>
 	);
 }
