+++ src/worldwise.code-workspace
@@ -1,0 +1,9 @@
+{
+	"folders": [
+		{
+			"name": "worldwise",
+			"path": ".."
+		}
+	],
+	"settings": {}
+}
